Moved profile logic to new file
