+++ components/ui/screen-header/index.tsx
@@ -5,7 +5,7 @@ import Colors from "@/shared-uis/constants/Colors";
 import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-native-fontawesome";
 import { useTheme } from "@react-navigation/native";
-import { Platform, Pressable } from "react-native";
+import { Platform, Pressable, View as RNView } from "react-native";
 import { Appbar } from "react-native-paper";
 
 interface ScreenHeaderProps {
@@ -13,6 +13,8 @@ interface ScreenHeaderProps {
     title: string;
     rightActionButton?: React.ReactNode;
     rightAction?: boolean;
+    /** When false, hides the back affordance (e.g. root tab screens). */
+    showBack?: boolean;
 }
 
 const ScreenHeader: React.FC<ScreenHeaderProps> = ({
@@ -20,6 +22,7 @@ const ScreenHeader: React.FC<ScreenHeaderProps> = ({
     title,
     rightActionButton,
     rightAction = false,
+    showBack = true,
 }) => {
     const theme = useTheme();
     const { user } = useAuthContext();
@@ -45,30 +48,40 @@ const ScreenHeader: React.FC<ScreenHeaderProps> = ({
             }}
             statusBarHeight={0}
         >
-            <Pressable onPress={handleAction}>
-                <View
-                    style={{
-                        marginTop: 2,
-                        marginLeft: 16,
-                    }}
-                    lightColor={Colors(theme).transparent}
-                    darkColor={Colors(theme).transparent}
-                >
-                    <FontAwesomeIcon
-                        icon={faArrowLeft}
-                        size={20}
-                        color={Colors(theme).text}
+            {showBack ? (
+                <Pressable onPress={handleAction}>
+                    <View
                         style={{
-                            alignItems: "center",
+                            marginTop: 2,
+                            marginLeft: 16,
                         }}
-                    />
-                </View>
-            </Pressable>
+                        lightColor={Colors(theme).transparent}
+                        darkColor={Colors(theme).transparent}
+                    >
+                        <FontAwesomeIcon
+                            icon={faArrowLeft}
+                            size={20}
+                            color={Colors(theme).text}
+                            style={{
+                                alignItems: "center",
+                            }}
+                        />
+                    </View>
+                </Pressable>
+            ) : (
+                <RNView style={{ width: 16 }} />
+            )}
 
             <Appbar.Content
                 style={{
                     flex: 1,
-                    marginLeft: Platform.OS === "web" ? 32 : 16,
+                    marginLeft: showBack
+                        ? Platform.OS === "web"
+                            ? 32
+                            : 16
+                        : Platform.OS === "web"
+                          ? 16
+                          : 8,
                 }}
                 color={Colors(theme).text}
                 title={title}
